[FEATURE][TAS-75] Added home page featured events
