+++ desparchado/views/__init__.py
@@ -1,0 +1,2 @@
+from .home import HomeView
+from .rss import RssSiteEventsFeed, SocialNetworksRssSiteEventsFeed, AtomSiteEventsFeed
