+++ desparchado/views/rss.py
@@ -1,4 +1,3 @@
-from django.views.generic import TemplateView
 from django.contrib.syndication.views import Feed
 from django.urls import reverse
 from django.utils.feedgenerator import Atom1Feed
@@ -6,29 +5,7 @@ from django.utils import timezone
 
 from events.models import Event
 from events.models import SocialNetworkPost
-from blog.models import Post
 
-
-class HomeView(TemplateView):
-    template_name = 'desparchado/index.html'
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        future_events_limit = 18
-        past_events_limit = 18
-        blog_posts_limit = 6
-
-        events = Event.objects.published().future().order_by('?').all()[:future_events_limit]
-
-        context['events'] = events
-        if events.count() <= 6:
-            context['past_events'] = \
-                Event.objects.published().past().order_by('-event_date').all()[:past_events_limit]
-        else:
-            context['past_events'] = []
-
-        context['blog_posts'] = Post.objects.published().order_by('?').all()[:blog_posts_limit]
-        return context
 
 
 class RssSiteEventsFeed(Feed):
